IP must be included in a folder
According to the specification, the IP must be enclosed in a folder - the folder name corresponding to the package name, for example. For convenience, the SIP-AIP conversion can also process SIPs where the IP's content is not enclosed in a folder, but directly put in the root of the package file.